+++ CHANGELOG.md
@@ -13,6 +13,7 @@ _When adding new entries to the changelog, please include issue/PR numbers where
 - Fixes a bug where features are written to the working copy without their CRS identifier during `kart resolve`. [#840](https://github.com/koordinates/kart/issues/840)
 - Drop CI testing of amazonlinux:2, replace with amazonlinux:2023. [#838](https://github.com/koordinates/kart/pull/838)
 - Point clouds (or rasters): during import, require the user to choose if they want to make a "cloud-optimized" dataset or a "preserve-format" dataset. [#842](https://github.com/koordinates/kart/issues/842)
+- Point clouds (or rasters): require `--no-convert-to-dataset-format` to add new tiles to a dataset that can be committed as-is but only if the dataset's format is changed. Note `--convert-to-dataset-format` works as before. [#842](https://github.com/koordinates/kart/issues/842)
 
 ## 0.12.3